+++ src/models/brick.lua
@@ -4,10 +4,11 @@ local brick = {
     width = 60,           -- Default width
     height = 20,          -- Default height
     shouldRemove = false, -- Boolean flag for pending removal
+    sprite = {},
 }
 
 -- Initialize brick position and body
-function brick.new(world, x, y)
+function brick.new(world, x, y, sprite)
     local b = {}
 
     b.width = brick.width
@@ -17,22 +18,19 @@ function brick.new(world, x, y)
     b.fixture = love.physics.newFixture(b.body, b.shape)
     b.fixture:setUserData({ type = "brick", obj = b })
 
+    b.sprite = sprite
     return b
 end
-
 
 -- Draw brick to screen
 function brick.draw(b)
     love.graphics.setColor(colors.BRICK)
+
     local x, y = b.body:getPosition()
-    love.graphics.rectangle(
-        "fill",
-        x - b.width / 2,
-        y - b.height / 2,
-        b.width,
-        b.height,
-        2, 2
-    )
+    local originX = b.width / 2
+    local originY = b.height / 2
+
+    love.graphics.draw(b.sprite, x, y, 0, 1, 1, originX, originY)
 end
 
 return brick