+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.2
+
+- Updated game sprites
+- Fixed ball speed not reset when switch back to main screen
+- Increase ball speed on brick destroyed
+
 ## 0.0.1
 
 This is just the barebone working prototype of the project, take me a good while to figuring out how to work with Love2D physics though.